# [SPARK-23963] [SQL] Properly handle large number of columns in query on text-based Hive table