@vanzin 2018-04-12
LGTM but let's see if someone more familiar with this code has comments.