@Tagar 2018-04-18
@gatorsmile could you please backport this to a Spark 2.2 branch as well?
This PR gives 24x improvement on 6000 columns as @bersprockets discovered, so I think this 1-line change should be fairly safely applied to Spark 2.2 as well. We see the same performance degradation on wide dataframes in Spark 2.2 as well. Thanks both!